Converting divisions between integers to an integer

+++ xmipp3/tests/test_protocols_xmipp_movies.py
@@ -464,10 +464,10 @@ class TestEstimateGain(BaseTest):
 
         frameImg = ih.createImage()
         xdim, ydim, zdim, ndim = ih.getDimensions(inputFn)
-        n = max(zdim, ndim) / 2  # also half of the frames
+        n = int(max(zdim, ndim) / 2)  # also half of the frames
         print("Scaling movie: %s -> %s" % (inputFn, outputFn))
         pwutils.cleanPath(outputFn)
-        for i in range(1, int(n+1)):
+        for i in range(1, n+1):
             frameImg.read((i, inputFn))
             frameImg.scale(int(xdim/2), int(ydim/2))
             frameImg.write((i, outputFn))
